check existance before deleting log file

+++ Casks/scripts/uninstall.sh
@@ -4,4 +4,6 @@ export PATH=/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/Library/Apple/usr/bin
 if [ -f /usr/local/bin/netclient ]; then
         netclient uninstall
 fi
-rm /var/log/com.gravitl.netclient.log
+if [ -f /usr/local/bin/netclient ]; then
+    rm /var/log/com.gravitl.netclient.log
+fi
